refactor: domain 패키지 구조 변경 및 이름 변경
- 변경 사항 1: domain->entity 파일 이름 변경 및 패키지 경로 이동
